shim comparison using sha256 now, closes #57
convert from md5 to sha256 for the shim comparison so it can be operated in a FIPS environment. This was not a cryptographic comparison, but unfortunately fast non-cryptographic hashes like xxhash or murmur are not yet available in (especially old) versions of python, but sha256 can be safely used everywhere.